Backend/requirements.txt
asgiref==3.8.1
autopep8==2.0.0
axios==0.4.0
beautifulsoup4==4.12.3
black==22.10.0
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
confection==0.0.4
cymem==2.0.7
defusedxml==0.7.1
distlib==0.3.9
Django==4.2.15
django-bootstrap-v5==1.0.11
django-cors-headers==4.4.0
django-filter==24.3
django-restframework==0.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
filelock==3.16.1
httpie==3.2.3
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
langcodes==3.3.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.2
mdurl==0.1.2
multidict==6.0.5
murmurhash==1.0.9
mypy==0.991
mypy-extensions==0.4.3
nltk==3.8.1
numpy==1.24.2
opencv-python==4.10.0.84
packaging==23.1
pathspec==0.10.2
pathy==0.10.1
pillow==10.4.0
platformdirs==4.3.6
preshed==3.0.8
pycodestyle==2.10.0
pydantic==1.10.7
Pygments==2.18.0
PyJWT==2.9.0
PySocks==1.7.1
pytesseract==0.3.10
regex==2023.3.23
requests==2.28.2
requests-toolbelt==1.0.0
rich==13.7.1
six==1.17.0
smart-open==6.3.0
soupsieve==2.5
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlparse==0.5.1
srsly==2.4.6
thinc==8.1.9
tomli==2.0.1
tqdm==4.65.0
typer==0.7.0
typing_extensions==4.4.0
tzdata==2024.1
urllib3==1.26.15
virtualenv==20.27.1
wasabi==1.1.1
